cleanup, rename files
