--- tests/jest-setup-after-env.ts
@@ -1,18 +1,0 @@
-import { vi, expect } from 'vitest';
-import * as matchers from 'jest-extended';
-
-expect.extend(matchers);
-
-Object.defineProperty(window, 'matchMedia', {
-  writable: true,
-  value: vi.fn().mockImplementation((query) => ({
-    matches: false,
-    media: query,
-    onchange: null,
-    addListener: vi.fn(), // Deprecated
-    removeListener: vi.fn(), // Deprecated
-    addEventListener: vi.fn(),
-    removeEventListener: vi.fn(),
-    dispatchEvent: vi.fn(),
-  })),
-});
